feat: add eslint config
