+++ .vscode/settings.json
@@ -1,0 +1,11 @@
+{
+    "eslint.enable": true,
+    "eslint.autoFixOnSave": true,
+    "eslint.options": {
+        "configFile": "/FrontEnd/virtual-keyboard/.eslintrc.js"
+    },
+    "editor.codeActionsOnSave": {
+        "source.fixAll.eslint": true
+    }
+
+}
