+++ index.html
@@ -7,6 +7,6 @@
     <title>Keyboard</title>
 </head>
 <body>
-    <script src="src/script.js" type="module"></script>
+    <script src="index.js" type="module"></script>
 </body>
 </html>
